The project upload
Input uses ‘*’ as AND, ‘+’ as OR, ‘!’ for negation, ‘>’ for Conditional statement, ‘==’ as Biconditional and ‘!=’ as XOR. Variables can go under any name consisting from characters and digits, but ‘{}’ and ‘[]’ are reserved and should not be used. None of the variables should fully include other’s name (Examples: a0 and a1 are valid, a and a1 / a1 and a10 are not).